Refactor some FSD layers
