+++ .nvmrc
@@ -1,1 +1,1 @@
-v18.12.1
+v18.16.0
