+++ package.json
@@ -69,7 +69,7 @@
     "prettier": "2.8.7",
     "rollup-plugin-visualizer": "5.9.0",
     "storybook": "7.0.9",
-    "storybook-addon-react-router-v6": "1.3.6--canary.ffa6de3.0",
+    "storybook-addon-react-router-v6": "1.0.1",
     "stylelint": "15.5.0",
     "stylelint-config-standard": "33.0.0",
     "ts-to-zod": "3.0.0",
