+++ src/widgets/CategoryPopularList/ui/CategoryPopularList/CategoryPopularList.module.css
