+++ src/widgets/LayoutHeader/ui/LayoutHeader/LayoutHeader.module.css
@@ -14,8 +14,3 @@
   gap: 16px;
 }
 
-.search {
-  width: 24px;
-  height: 24px;
-  background-image: url('/search.svg');
-}
