+++ src/widgets/LayoutHeader/ui/Logo/Logo.module.css
@@ -1,4 +1,4 @@
-@import url("../../shared/breakpoints.css");
+@import url("../../../../shared/breakpoints.css");
 
 .root {
   width: 180px;
