+++ src/widgets/LayoutProfileCard/LayoutProfileCard.tsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom'
 import { selectTotalQuantity } from '@/entities/cart'
 import { selectIsAuthorized, useMeQuery } from '@/entities/session'
 import { selectProductIdsInWishlist } from '@/entities/wishlist'
-import { LogoutButton } from '@/features/authentication/Logout'
+import { LogoutButton } from '@/features/authentication/logout'
 import { useFeatureSlicedDebug } from '@/shared/lib'
 import { useAppSelector } from '@/shared/model'
 import { Icon } from '@/shared/ui'
