--- src/widgets/Logo/Logo.tsx
@@ -1,14 +1,0 @@
-import { Link } from 'react-router-dom'
-import { useFeatureSlicedDebug } from '@/shared/lib'
-import css from './Logo.module.css'
-
-export function Logo() {
-  const { rootAttributes } =
-    useFeatureSlicedDebug<HTMLAnchorElement>('widget/Logo')
-
-  return (
-    <Link to={'/'} {...rootAttributes}>
-      <div className={css.root}></div>
-    </Link>
-  )
-}
