+++ src/widgets/ProductList/ProductList.tsx
@@ -2,7 +2,7 @@ import cn from 'classnames'
 import { type ReactNode } from 'react'
 import { ProductCard, type Product, type ProductId } from '@/entities/product'
 import { selectIsAuthorized } from '@/entities/session'
-import { AddToWishlistIcon } from '@/features/wishlist/AddToWishlist'
+import { AddToWishlistIcon } from '@/features/wishlist/addToWishlist'
 import { useFeatureSlicedDebug } from '@/shared/lib'
 import { useAppSelector } from '@/shared/model'
 import css from './ProductList.module.css'
